+++ pint_xarray/tests/test_accessors.py
@@ -247,7 +247,12 @@ class TestQuantifyDataSet:
 
     def test_error_on_nonsense_units(self, example_unitless_ds):
         ds = example_unitless_ds
-        with pytest.raises(UndefinedUnitError):
+        with pytest.raises(ValueError):
+            ds.pint.quantify(units={"users": "aecjhbav"})
+
+    def test_error_indicates_problematic_variable(self, example_unitless_ds):
+        ds = example_unitless_ds
+        with pytest.raises(ValueError, match="users"):
             ds.pint.quantify(units={"users": "aecjhbav"})